requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
discord.py==2.5.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.1.0
playwright==1.50.0
propcache==0.3.0
pyee==12.1.1
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
